[ALPS07154784] tf-a: fix coverity issues on tfa_debug.c
Fix coverity issues on tfa_debug.c

MTK-Commit-Id: 18c6b7b3e46750a58e4080e35f93e8f07bfc0ced

CR-Id: ALPS07154784
Feature: TF-A

+++ drivers/misc/mediatek/atf/tfa_debug.c
@@ -135,7 +135,7 @@ static int generic_open(struct inode *inode, struct file *file)
 		GFP_KERNEL);
 	if (IS_ERR(inst_p)) {
 		file->private_data = NULL;
-		pr_notice("Fail to kmalloc:%d\n", PTR_ERR(inst_p));
+		pr_notice("Fail to kmalloc:%ld\n", PTR_ERR(inst_p));
 		return -ENOMEM;
 	}
 	init_waitqueue_head(&inst_p->waitq);
@@ -167,7 +167,7 @@ static int raw_open(struct inode *inode, struct file *file)
 
 static int raw_release(struct inode *node, struct file *file)
 {
-	pr_info("%s:vaddr:0x%llx\n", __func__, file->private_data);
+	pr_info("%s:vaddr:0x%llx\n", __func__, (uint64_t)file->private_data);
 	if (file->private_data) {
 		struct tfa_debug_instance *inst_p = file->private_data;
 
@@ -199,7 +199,7 @@ static ssize_t raw_read(struct file *file,
 	/* Update the read pos */
 	inst_p->read_offset += copy_len;
 	pr_info("vaddr:0x%llx, count:%zu, read offset:%zu\n",
-		inst_p->vaddr, count, inst_p->read_offset);
+		(uint64_t)inst_p->vaddr, count, inst_p->read_offset);
 	return copy_len;
 }
 
@@ -259,7 +259,7 @@ static int getc_for_read(struct file *file, char *p)
 		return -1;
 	body = inst_p->vaddr;
 	if (inst_p->read_offset >= inst_p->debug_buf_size) {
-		pr_notice("%s read offset is invalid, :%u\n",
+		pr_notice("%s read offset is invalid, :%lu\n",
 			__func__, inst_p->read_offset);
 		inst_p->read_offset = 0;
 	}
@@ -407,11 +407,11 @@ static int lookup_reserved_memory(void)
 	/* remap reserved memory as cacheale */
 	info.vaddr = ioremap(info.debug_buf_paddr, info.total_size);
 	if (IS_ERR(info.vaddr)) {
-		pr_notice("Fail to remap debug buf vaddr:%d\n",
+		pr_notice("Fail to remap debug buf vaddr:%ld\n",
 			PTR_ERR(info.vaddr));
 		return -ENOMEM;
 	}
-	pr_info("debug buf vaddr:0x%llx\n", info.vaddr);
+	pr_info("debug buf vaddr:0x%llx\n", (uint64_t)info.vaddr);
 	return 0;
 }
